+++ deepcave/runs/__init__.py
@@ -40,6 +40,7 @@ class Status(IntEnum):
     CRASHED = 4
     ABORTED = 5
     RUNNING = 6
+    NOTFOUND = 7
 
 
 @dataclass
@@ -59,7 +60,7 @@ class Trial:
         assert isinstance(self.status, Status)
 
     def get_key(self) -> Tuple[int, int]:
-        return self.config_id, self.budget  # noqa
+        return AbstractRun.get_trial_key(self.config_id, self.budget)
 
     def to_json(self) -> List[Any]:
         return [
@@ -100,14 +101,14 @@ class AbstractRun(ABC):
         """
         Hash of the current run. If hash changes, cache has to be cleared. This ensures that
         the cache always holds the latest results of the run.
-        
+
         Returns
         -------
         str
             Hash of the run.
         """
         pass
-    
+
     @property
     @abstractmethod
     def id(self) -> str:
@@ -121,6 +122,16 @@ class AbstractRun(ABC):
             Hash of the run.
         """
         pass
+
+    @staticmethod
+    def get_trial_key(config_id, budget):
+        return (config_id, budget)
+
+    def get_trial(self, trial_key) -> Optional[Trial]:
+        if trial_key not in self.trial_keys:
+            return None
+
+        return self.history[self.trial_keys[trial_key]]
 
     def get_trials(self) -> Iterator[Trial]:
         yield from self.history
@@ -206,14 +217,18 @@ class AbstractRun(ABC):
         return [obj["name"] for obj in self.get_objectives()]
 
     def get_configs(self, budget=None) -> List:
+        config_ids = []
         configs = []
         for trial in self.history:
             if budget is not None:
                 if budget != trial.budget:
                     continue
 
-            config = self.configs[trial.config_id]
-            configs += [config]
+            if (config_id := trial.config_id) not in config_ids:
+                config = self.configs[config_id]
+
+                config_ids += [config_id]
+                configs += [config]
 
         return configs
 
@@ -227,6 +242,9 @@ class AbstractRun(ABC):
                 return id
 
         return None
+
+    def get_num_configs(self, budget=None) -> int:
+        return len(self.get_configs(budget=budget))
 
     def get_budget(self, id: int) -> float:
         return self.meta["budgets"][id]