+++ deepcave/runs/converters/bohb.py
@@ -1,3 +1,4 @@
+import os
 from pathlib import Path
 
 from deepcave.runs import Status
@@ -18,6 +19,10 @@ class BOHBRun(Run):
         """
         # Use hash of results.json as id
         return file_to_hash(self.path / "results.json")
+
+    @property
+    def latest_change(self) -> int:
+        return Path(self.path / "results.json").stat().st_mtime
 
     @classmethod
     def from_path(cls, path: Path) -> "BOHBRun":