+++ deepcave/runs/grouped_run.py
@@ -96,6 +96,15 @@ class GroupedRun(AbstractRun):
         return string_to_hash(f"{self.prefix}:{self.name}")
 
     @property
+    def latest_change(self) -> int:
+        latest_change = 0
+        for run in self.runs:
+            if run.latest_change > latest_change:
+                latest_change = run.latest_change
+
+        return latest_change
+
+    @property
     def run_paths(self) -> List[str]:
         return [str(run.path) for run in self.runs]